update css, update project structure
