+++ front/src/App.css
@@ -1,0 +1,13 @@
+.app-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  gap: 1.5rem;
+}
+
+.app-title {
+  font-size: 2rem;
+  font-weight: bold;
+}
