+++ front/src/components/Button.css
@@ -1,0 +1,21 @@
+.btn {
+  color: white;
+  padding: 0.5rem 1rem;
+  border: none;
+  border-radius: 0.375rem;
+  cursor: pointer;
+  font-size: 1rem;
+}
+
+.btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.btn-roll {
+  background-color: #3b82f6;
+}
+
+.btn-cashout {
+  background-color: #facc15;
+}
