+++ front/src/index.css
@@ -1,5 +1,10 @@
-/*
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-*/
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  font-family: sans-serif;
+  background-color: #d1fae5;
+}
