+++ front/src/main.jsx
@@ -1,7 +1,8 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.jsx'
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import './index.css';
+
+import App from './App.jsx';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
